refactor(Store): using macro for impl module boilerplate
